requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.36
botocore==1.38.36
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
npm==0.1.1
optional-django==0.1.0
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==1.7.1
pytest==8.4.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
